backend-python/requirements.txt
Django==5.1.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.4.0
redis==5.2.1
Pillow==11.0.0
django-filter==24.3
drf-spectacular==0.28.0
boto3==1.35.80
django-storages==1.14.5
django-import-export==4.3.3
pandas==2.2.3
openpyxl==3.1.5
django-modeltranslation==0.19.9
gunicorn==23.0.0
whitenoise==6.8.2
icalendar==6.1.0
pytz==2024.2
django-redis==5.4.0
django-ratelimit==4.1.0
django-csp==3.8

# Dev & security tooling (optional)
bandit==1.8.0
pip-audit==2.7.3
dj-database-url==2.1.0

# Needed for integration tests that perform real HTTP requests
requests==2.31.0

# Supabase client for Storage
supabase==1.0.0
